timerow: Change design to pill type.
The aim of this and the child commits is to change the behavior of
the timerow to be persistent. The design has to be changed to make
the new behavior viable. The pill design seems to be the perfect
choice for our usecase, which is supported by the fact that it is
what the other popular messaging applications use.

To make this possible the decorations on the left and the right of
the timerow were removed and a central element was added to make the
timerow. A span was introduced to seperate the timerow container
from the content of the timerow itself, and for ensuring that the
margins and the background colors worked properly (only for the
text, rather than for the whole container).