timerow: Add persistent timerow element.
On a recent discussion on chat.zulip.org it was determined that a
good way to go about showing the date consistently in the WebView
was to just copy the sticky timerow element from WhatsApp. That
implementation keeps a timerow, which from now will be called
`date-pill` because of its true nature, is always on top of the
screen, and as the other date-pills whiz by it, its content gets
updated.

This commit just adds that sticky date-pill element to the
screen. It gets its `top` property set by the `hasRecipientHeaders`
property, which is sent to the `css.js` default export. Since it is
currently empty it will not show because of :empty pseudoelement
added to `css.js`. The persistent date-pill is very slightly raised
in order to not appear janky, and it's `z-index` has been raised to
above the headers.

NOTE: The timerow/date-pill now appears below a header instead of
above it when they appear together. The tests have been amended with
respect to that, and an additional line is added to each test which
tests for a header object.