date-pill: Make date-pill-sticky appear below the sticky recipient header.
This is an experimental commit to make the date=pill appear below
the sticky headers while appearing below the ones which haven't
stuck yet. We make a mask over the webview below the headers and
extending upto the end of the screen. This mask is set with
`overflow: hidden` and the transition works as it did before,
essentially hiding the date pill above a certain amount.